break-out time-filter states into dedicated file: src/recoil/time-filters.js
